Use rewrites to improve performance
Rewrites save us ~15ms per request because (I assume) they just shunt
the traffic to the next route rather than setting up proxy buffers and
so forth. Anyway: they’re faster, so we should use them.

These appear to be the only places we can use rewrites: small_light has
to be the first URI called in the chain, so we have to proxy to it.